refactore the X25519 keys
