+++ src/asymmetric_crypto/salsa_sealbox/mod.rs
@@ -1,7 +1,7 @@
 mod ecies;
-mod keypair;
+mod private_key;
+mod public_key;
 
 pub use ecies::EciesSalsaSealBox;
-pub use keypair::{
-    X25519PrivateKey, X25519PublicKey, X25519_PRIVATE_KEY_LENGTH, X25519_PUBLIC_KEY_LENGTH,
-};
+pub use private_key::{X25519PrivateKey, X25519_PRIVATE_KEY_LENGTH};
+pub use public_key::{X25519PublicKey, X25519_PUBLIC_KEY_LENGTH};
